move IsAllowedNamespaces
